chore: split internal code and code from Go
