+++ .golangci.yml
@@ -221,11 +221,7 @@ issues:
 
   exclude-dirs:
     - test/testdata_etc # test files
-    - internal/cache # extracted from Go code
-    - internal/robustio # extracted from Go code
-    - internal/mmap # extracted from Go code
-    - internal/quoted # extracted from Go code
-    - internal/testenv # extracted from Go code
+    - internal/go # extracted from Go code
 
 run:
   timeout: 5m
