+++ internal/go/LICENSE
