+++ internal/go/cache/hash.go
