+++ internal/go/cache/hash_gcil.go
