+++ internal/go/cache/hash_test.go
