+++ internal/go/mmap/mmap_other.go
