+++ internal/go/mmap/readme.md
