+++ internal/go/quoted/quoted.go
