+++ internal/go/robustio/readme.md
